`image_background`+`bevy_api_testing` examples
MV `bevy_api_testing.rs` > `sprite_and_ui_clickable.rs`
