+++ examples/image_background.rs
@@ -1,5 +1,6 @@
 use bevy::prelude::*;
 use bevy_cosmic_edit::*;
+use util::{bevy_color_to_cosmic, change_active_editor_ui, deselect_editor_on_esc};
 
 fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
     commands.spawn(Camera2dBundle::default());
@@ -30,17 +31,6 @@ fn setup(mut commands: Commands, asset_server: Res<AssetServer>) {
             ..default()
         })
         .insert(CosmicSource(editor));
-
-    commands.insert_resource(FocusedWidget(Some(editor)));
-}
-
-pub fn bevy_color_to_cosmic(color: bevy::prelude::Color) -> CosmicColor {
-    CosmicColor::rgba(
-        (color.r() * 255.) as u8,
-        (color.g() * 255.) as u8,
-        (color.b() * 255.) as u8,
-        (color.a() * 255.) as u8,
-    )
 }
 
 fn main() {
@@ -48,5 +38,6 @@ fn main() {
         .add_plugins(DefaultPlugins)
         .add_plugins(CosmicEditPlugin::default())
         .add_systems(Startup, setup)
+        .add_systems(Update, (change_active_editor_ui, deselect_editor_on_esc))
         .run();
 }
