+++ util/src/lib.rs
@@ -73,3 +73,12 @@ pub fn print_editor_text(
         info!("Widget text: {:?}", current_text);
     }
 }
+
+pub fn bevy_color_to_cosmic(color: bevy::prelude::Color) -> CosmicColor {
+    CosmicColor::rgba(
+        (color.r() * 255.) as u8,
+        (color.g() * 255.) as u8,
+        (color.b() * 255.) as u8,
+        (color.a() * 255.) as u8,
+    )
+}
